chore: #277 - enum 분리
