+++ Spoony-iOS/Spoony-iOS/Network/DependencyKey/HomeServiceKey.swift
@@ -1,0 +1,12 @@
+//
+//  HomeServiceKey.swift
+//  Spoony-iOS
+//
+//  Created by 최안용 on 5/4/25.
+//
+
+import Dependencies
+
+enum HomeServiceKey: DependencyKey {
+    static let liveValue: HomeServiceType = DefaultHomeService()
+}
