+++ Spoony-iOS/Spoony-iOS/Network/DependencyKey/RegisterServiceKey.swift
@@ -1,0 +1,12 @@
+//
+//  RegisterServiceKey.swift
+//  Spoony-iOS
+//
+//  Created by 최안용 on 5/4/25.
+//
+
+import Dependencies
+
+enum RegisterServiceKey: DependencyKey {
+    static let liveValue: RegisterServiceType = RegisterService()
+}
